pyproject.toml
[project]
name = "with-bm25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-groq>=1.1.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pymupdf>=1.27.2.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "streamlit>=1.56.0",
]
